// frontend/src/openapi/models/TrainingMonitoring.ts
export interface TrainingMonitoring {
    /**
     *
     * @type {string}
     * @memberof TrainingMonitoring
     */
    jobId: string;
    /**
     *
     * @type {any}
     * @memberof TrainingMonitoring
     */
    policy: any | null;
    /**
     *
     * @type {string}
     * @memberof TrainingMonitoring
     */
    status: string;
    /**
     *
     * @type {any}
     * @memberof TrainingMonitoring
     */
    current: any | null;
    /**
     *
     * @type {any}
     * @memberof TrainingMonitoring
     */
    probes: any | null;
    /**
     *
     * @type {number}
     * @memberof TrainingMonitoring
     */
    count: number;
    /**
     *
     * @type {number}
     * @memberof TrainingMonitoring
     */
    nextOffset: number | null;
    /**
     *
     * @type {Array<TrainingCheck>}
     * @memberof TrainingMonitoring
     */
    checks: Array<TrainingCheck>;
    /**
     *
     * @type {Array<TrainingCheckpoint>}
     * @memberof TrainingMonitoring
     */
    checkpoints: Array<TrainingCheckpoint>;
    /**
     *
     * @type {Array<string>}
     * @memberof TrainingMonitoring
     */
    limitations: Array<string>;
}